requirements.txt
PyYAML>=6.0
robotframework>=4.13
robotframework-assertion-engine>=0.5.1
robotframework-browser>=11.4.0
robotframework-datadriver>=1.6.0
robotframework-pabot>=2.1.0
robotframework-pythonlibcore>=3.0.0
robotframework-requests>=0.9.2
robotframework-robocop>=1.12.0
robotframework-seleniumlibrary>=6.0.0
robotframework-stacktrace>=0.4.1
robotframework-tidy>=1.6.2
pre-commit>=2.17.0
